examples/api_integration_assistant/requirements.txt
firecrawl-py>=1.9.0
pydantic>=2.0.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
aiohttp>=3.8.0
pytest>=7.0.0
pytest-asyncio>=0.18.0